refactor delete
